wazuh_integration/requirements.txt
aiohttp
python-dotenv
opensearch-py
pydantic
requests
loguru
